+++ jid/CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 
   * Breaking
     - serde: Jid is now using untagged enum representation (#66)
+  * Additions
+    - From<&Jid> is now implemented for String (#69)
 
 Version 0.9.3, release 2022-03-07:
   * Updates